[IMP][SI-770] Remove the redundant code and add some description
