+++ SI/mps_connector/__init__.py
@@ -5,6 +5,7 @@ from . import utils
 
 from odoo import api, SUPERUSER_ID
 
-def init_forecast_result_from_mps_data(cr, registry):
+
+def setup_mps_connector(cr, registry):
     env = api.Environment(cr, SUPERUSER_ID, {})
-    env['mps.connector'].sudo().init_forecast_result_from_mps_data()
+    env['mrp.production.schedule'].sudo().init_forecast_result_from_mps_data()
